"Physics"
!!CAUTION!!
This is quick and dirty since the professor said that this Task will be
the last and  will not contribute to our grade.
!!CAUTION!!

Add special treatment for the spheres and a bounding box
Fix vector arithmetic
Fix matrix arithmetic
Add function to transfer screen coordinates to world coordinates

+++ include/animated_cube.hpp
@@ -16,18 +16,18 @@ class AnimatedCube : public Cube {
 
   State state_;
 
-  float open_angel_;
-  float closed_angel_;
-  float current_angle_;
-  float animation_time_;
+  double open_angel_;
+  double closed_angel_;
+  double current_angle_;
+  double animation_time_;
 
   void calculate_current_angle();
 
 public:
 
   AnimatedCube();
-  AnimatedCube(float origin_x, float origin_y, float origin_z);
-  AnimatedCube(float origin_x, float origin_y, float origin_z,
+  AnimatedCube(double origin_x, double origin_y, double origin_z);
+  AnimatedCube(double origin_x, double origin_y, double origin_z,
                unsigned char colors);
   ~AnimatedCube();
